update tf_3dunet_brats to newer convention and add general plan
