+++ openfl-workspace/tf_3dunet_brats/plan/defaults
@@ -1,0 +1,1 @@
+../../workspace/plan/defaults
