--- openfl-workspace/tf_3dunet_brats/plan/defaults/compression_pipeline.yaml
@@ -1,1 +1,0 @@
-template: openfl.pipelines.NoCompressionPipeline
