--- openfl-workspace/tf_3dunet_brats/plan/defaults/data_loader.yaml
@@ -1,1 +1,0 @@
-template: openfl.federated.DataLoader
