--- openfl-workspace/tf_3dunet_brats/plan/defaults/defaults
@@ -1,2 +1,0 @@
-../../workspace/plan/defaults
-
