+++ openfl-workspace/tf_3dunet_brats/src/define_model.py
@@ -8,11 +8,16 @@ import tensorflow as tf
 
 def dice_coef(target, prediction, axis=(1, 2, 3), smooth=0.0001):
     """
-    Sorenson Dice.
+    Calculate the Sorenson-Dice coefficient.
 
-    Returns
-    -------
-    dice coefficient (float)
+    Args:
+        target (tf.Tensor): The ground truth binary labels.
+        prediction (tf.Tensor): The predicted binary labels, rounded to 0 or 1.
+        axis (tuple, optional): The axes along which to compute the coefficient, typically the spatial dimensions.
+        smooth (float, optional): A small constant added to numerator and denominator for numerical stability.
+
+    Returns:
+        tf.Tensor: The mean Dice coefficient over the batch.
     """
     prediction = tf.round(prediction)  # Round to 0 or 1
 
@@ -27,13 +32,18 @@ def dice_coef(target, prediction, axis=(1, 2, 3), smooth=0.0001):
 
 def soft_dice_coef(target, prediction, axis=(1, 2, 3), smooth=0.0001):
     """
-    Soft Sorenson Dice.
+    Calculate the soft Sorenson-Dice coefficient.
 
     Does not round the predictions to either 0 or 1.
 
-    Returns
-    -------
-    soft dice coefficient (float)
+    Args:
+        target (tf.Tensor): The ground truth binary labels.
+        prediction (tf.Tensor): The predicted probabilities.
+        axis (tuple, optional): The axes along which to compute the coefficient, typically the spatial dimensions.
+        smooth (float, optional): A small constant added to numerator and denominator for numerical stability.
+
+    Returns:
+        tf.Tensor: The mean soft Dice coefficient over the batch.
     """
     intersection = tf.reduce_sum(target * prediction, axis=axis)
     union = tf.reduce_sum(target + prediction, axis=axis)
@@ -46,15 +56,20 @@ def soft_dice_coef(target, prediction, axis=(1, 2, 3), smooth=0.0001):
 
 def dice_loss(target, prediction, axis=(1, 2, 3), smooth=0.0001):
     """
-    Sorenson (Soft) Dice loss.
+    Calculate the (Soft) Sorenson-Dice loss.
 
     Using -log(Dice) as the loss since it is better behaved.
     Also, the log allows avoidance of the division which
     can help prevent underflow when the numbers are very small.
 
-    Returns
-    -------
-    dice loss (float)
+    Args:
+        target (tf.Tensor): The ground truth binary labels.
+        prediction (tf.Tensor): The predicted probabilities.
+        axis (tuple, optional): The axes along which to compute the loss, typically the spatial dimensions.
+        smooth (float, optional): A small constant added to numerator and denominator for numerical stability.
+
+    Returns:
+        tf.Tensor: The mean Dice loss over the batch.
     """
     intersection = tf.reduce_sum(prediction * target, axis=axis)
     p = tf.reduce_sum(prediction, axis=axis)
@@ -70,29 +85,29 @@ def build_model(input_shape,
                 n_cl_out=1,
                 use_upsampling=False,
                 dropout=0.2,
-                print_summary=True,
                 seed=816,
                 depth=5,
                 dropout_at=(2, 3),
                 initial_filters=16,
-                batch_norm=True,
-                **kwargs):
-    """Build the TensorFlow model.
+                batch_norm=True,):
+    """
+    Build and compile 3D UNet model.
 
     Args:
-        input_tensor: input shape ot the model
-        use_upsampling (bool): True = use bilinear interpolation;
-                            False = use transposed convolution (Default=False)
+        input_shape (List[int]): The shape of the data
         n_cl_out (int): Number of channels in output layer (Default=1)
+        use_upsampling (bool): True = use bilinear interpolation;
+            False = use transposed convolution (Default=False)
         dropout (float): Dropout percentage (Default=0.2)
-        print_summary (bool): True = print the model summary (Default = True)
         seed: random seed (Default=816)
         depth (int): Number of max pooling layers in encoder (Default=5)
-        dropout_at: Layers to perform dropout after (Default=[2,3])
-        initial_filters (int): Number of filters in first convolutional
-        layer (Default=16)
-        batch_norm (bool): True = use batch normalization (Default=True)
-        **kwargs: Additional parameters to pass to the function
+        dropout_at (List[int]): Layers to perform dropout after (Default=[2,3])
+        initial_filters (int): Number of filters in first convolutional layer (Default=16)
+        batch_norm (bool): Aply batch normalization (Default=True)
+
+    Returns:
+        keras.src.engine.functional.Functional
+            A compiled Keras model ready for training.
     """
     if (input_shape[0] % (2**depth)) > 0:
         raise ValueError(f'Crop dimension must be a multiple of 2^(depth of U-Net) = {2**depth}')
